Make recent contributions count more than old ones
Problem:

People who haven't contributed in a while
can still top the contributors list
if they have contributed a lot in the past.

This isn't great – it means that the founder of a project
is likely to stay at the top of the list for a long time,
even if they have abandoned the project for a while.

Solution:

Weight contributions based on their recency.
Each contribution degrades exponentially over time,
giving new contributors a chance to rise up on the leaderboard.