Set up OAuth with GitHub
## Problem:

In order to take administrator actions on a repository,
we need to authenticate with a GitHub account
that has access to the repository.

Right now we authenticate with the developer's credentials,
stored in a `.netrc` file.
This solution only works for one user,
and is not secure to use outside of development.

## Solution:

Use the [omniauth-github] gem to authenticate the user
based on their GitHub account.

Once the user is authenticated,
we can use their access token to perform admin actions on their behalf.